pyproject.toml
[tool.poetry]
name = "pochurl"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
tinydb = "^4.8.0"
google-cloud-firestore = "^2.16.0"
firebase-admin = "^6.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
